Update ABE_ADCDACPi.h
replace:
double read_adc_raw(int channel, int mode);
by:
int read_adc_raw(int channel, int mode);

+++ ADCDACPi/ABE_ADCDACPi.h
@@ -23,7 +23,7 @@ void close_dac();
 
 double read_adc_voltage(int channel, int mode);
 
-double read_adc_raw(int channel, int mode);
+int read_adc_raw(int channel, int mode);
 
 void set_adc_refvoltage(double ref);
